Create Count The Paths
Here, I used map instead of naive container inside container (I learnt some stl for this problem) as accessing values are easier in this.

This algorithm in worst case (for a complete graph) will do n! computations.
n! <= n^n
 So time complexity is O(n^n).

And 8! < 8^8 < 10^8 , so exponential time is also feasible in this case.