Update Useless paths
I have used bellman ford algorithm for this problem, since it is an undirected connected graph, we initially consider 2*m paths available.

If there is a path available from a to b, I put -1 in the adjacency matrix p for marking it.

Bellman ford algorithm says to do relaxation for all the edges available for (nodes-1) times to get the correct result.
Since, all the cost of paths are positive, we need not worry about negative weighted cycles for which this algorithm fails.


If cost for moving from a to b is lesser, both paths a to b & b to a will be relaxed, so it would be a useful path & we would subtract both of these from total paths available.
As it is an undirected graph & we were considering both paths x to y & y to x till now.
We must divide the result by 2 to get correct no. of useless paths.

The main condition we are implementing is the relaxation using 4 nested for loops,each traversing from 1 to n.

So, the time complexity of this algorithm is O(n^4).
(100)^4 = 10^8 computations by system, which is feasible.